Get SFX_Play working and optimize
Fixes #2

I introduced a regression (ugly crash) when optimizing the rom bank switch when there is a sound to play.

I also changed the args to come in on bc instead of al. This removed the need for any pushes or pops and thus optimized the speed quite a bit. Also, bc is consistent with the rest of the API.