Initialize instrument banks
The instrument banks are left uninitialized, although the instrument
themselves are initialized to point to dummy instruments.

If the sound system is run for 256 frames before any song or
sound effect is played, BGB raises an "accessing uninitialized WRAM"
exception when attempting to switch to the instrument bank.

This diff initializes the banks to point to the appropriate one for the
dummy instrument (i.e. the sound system bank).

This still triggers an MBC1 write exception in BGB if the engine is in
bank 0, but it should be harmless.